zoomin

+++ experiments/por-sim-rs/src/chain/fork_rules.rs
@@ -13,9 +13,9 @@ use ForkResult::*;
 
 pub trait ForkRules<B: BlockT> {
     // fn new() -> Self;
-    fn fork_measure(b_md: &BlockMD<B>) -> u128;
+    fn fork_measure(b_md: &BlockMD<B>) -> u64;
     // fn best_of<'a>(b1: (&'a B, &BlockMD<B>), b2: (&'a B, &BlockMD<B>)) -> ForkResult<'a, B>;
-    fn weight_of<'a, F: ForkRules<B>>(b: &B, chain: &impl ChainT<'a, B, F>) -> u128;
+    fn weight_of<'a, F: ForkRules<B>>(b: &B, chain: &impl ChainT<'a, B, F>) -> u64;
     // fn weight_of<'a>(b: &B, f: Box<impl Fn(u128) -> Option<BlockMD<B>>>) -> u128;
 
     fn best_of<'a>(b1: (&'a B, &BlockMD<B>), b2: (&'a B, &BlockMD<B>)) -> ForkResult<'a, B> {
@@ -40,8 +40,8 @@ pub struct HeaviestChain<B: BlockT> {
 }
 
 impl<B: BlockT> ForkRules<B> for LongestChain<B> {
-    fn fork_measure(b_md: &BlockMD<B>) -> u128 {
-        b_md.height as u128
+    fn fork_measure(b_md: &BlockMD<B>) -> u64 {
+        b_md.height as u64
     }
 
     // fn best_of<'a>(b1: (&'a B, &BlockMD<B>), b2: (&'a B, &BlockMD<B>)) -> ForkResult<'a, B> {
@@ -55,8 +55,8 @@ impl<B: BlockT> ForkRules<B> for LongestChain<B> {
     //     }
     // }
 
-    fn weight_of<'a, F: ForkRules<B>>(b: &B, chain: &impl ChainT<'a, B, F>) -> u128 {
-        (chain.get_block_meta(b.prev()).unwrap().height + 1) as u128
+    fn weight_of<'a, F: ForkRules<B>>(b: &B, chain: &impl ChainT<'a, B, F>) -> u64 {
+        (chain.get_block_meta(b.prev()).unwrap().height + 1) as u64
     }
 
     // fn weight_of<'a>(b: &B, f: Box<impl Fn(u128) -> Option<BlockMD<B>>>) -> u128 {
@@ -65,11 +65,11 @@ impl<B: BlockT> ForkRules<B> for LongestChain<B> {
 }
 
 impl ForkRules<Block> for HeaviestChain<Block> {
-    fn fork_measure(b_md: &BlockMD<Block>) -> u128 {
+    fn fork_measure(b_md: &BlockMD<Block>) -> u64 {
         b_md.chain_weight
     }
 
-    fn weight_of<'a, F: ForkRules<Block>>(b: &Block, chain: &impl ChainT<'a, Block, F>) -> u128 {
+    fn weight_of<'a, F: ForkRules<Block>>(b: &Block, chain: &impl ChainT<'a, Block, F>) -> u64 {
         let p_id = b.prev();
         let p = chain.get_block(p_id).unwrap();
         let p_md = chain.get_block_meta(p_id).unwrap();
@@ -78,7 +78,7 @@ impl ForkRules<Block> for HeaviestChain<Block> {
 }
 
 impl ForkRules<DagBlock> for HeaviestChain<DagBlock> {
-    fn fork_measure(b_md: &BlockMD<DagBlock>) -> u128 {
+    fn fork_measure(b_md: &BlockMD<DagBlock>) -> u64 {
         b_md.chain_weight
     }
 
@@ -99,17 +99,17 @@ impl ForkRules<DagBlock> for HeaviestChain<DagBlock> {
     fn weight_of<'a, F: ForkRules<DagBlock>>(
         b: &DagBlock,
         chain: &impl ChainT<'a, DagBlock, F>,
-    ) -> u128 {
+    ) -> u64 {
         let (_lca, lca_intermediates) = chain.find_lca_and_intermediates(&b.parents).unwrap();
         let min_h = lca_intermediates.keys().min().unwrap();
         let max_h = lca_intermediates.keys().max().unwrap();
 
         let lca_info = &lca_intermediates[min_h].iter().cloned().collect::<Vec<_>>()[0];
         let base_chain_weight = lca_info.b_md.chain_weight;
-        let intermediate_weights: u128 = lca_intermediates
+        let intermediate_weights: u64 = lca_intermediates
             .iter()
             .filter(|(h, _)| *h != min_h)
-            .map::<u128, _>(|(_, infos)| infos.iter().map(|i| i.b_md.weight).sum())
+            .map::<u64, _>(|(_, infos)| infos.iter().map(|i| i.b_md.weight).sum())
             .sum();
 
         let some_p_info = lca_intermediates[max_h]
